+++ src/DataContainer/UserContainer.php
@@ -1,16 +1,12 @@
 <?php
-/**
- * Contao Open Source CMS
+
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
  *
- * Copyright (c) 2019 Heimrich & Hannot GmbH
- *
- * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
- * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ * @license LGPL-3.0-or-later
  */
 
-
 namespace HeimrichHannot\Email2UsernameBundle\DataContainer;
-
 
 use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
 use Contao\DataContainer;
@@ -21,14 +17,13 @@ use Contao\UserModel;
 class UserContainer
 {
     /**
-     * @var bool
-     */
-    private $enabled = true;
-
-    /**
      * @var ContaoFrameworkInterface
      */
     protected $framework;
+    /**
+     * @var bool
+     */
+    private $enabled = true;
 
     /**
      * UserContainer constructor.
@@ -44,7 +39,7 @@ class UserContainer
     public function onLoad(DC_Table $dcTable)
     {
         if (true === $this->enabled) {
-            $GLOBALS['TL_DCA']['tl_user']['fields']['username']['eval']['disabled']  = true;
+            $GLOBALS['TL_DCA']['tl_user']['fields']['username']['eval']['disabled'] = true;
             $GLOBALS['TL_DCA']['tl_user']['fields']['username']['eval']['mandatory'] = false;
         }
     }
@@ -57,36 +52,27 @@ class UserContainer
         $this->setUsernameFromEmail($dc);
     }
 
-
-
     /**
-     * Onsubmit callback
-     *
-     * @param DataContainer $dc
+     * Onsubmit callback.
      */
     protected function setUsernameFromEmail(DataContainer $dc)
     {
         /** @var UserModel $user */
-        if (null === $user && ($user = $this->framework->getAdapter(UserModel::class)->findByPk($dc->id)) === null) {
+        if (null === $user && null === ($user = $this->framework->getAdapter(UserModel::class)->findByPk($dc->id))) {
             return;
         }
 
-        if (null !== $dc->activeRecord)
-        {
+        if (null !== $dc->activeRecord) {
             $email = $dc->activeRecord->email;
-        }
-        else
-        {
-            if ($user instanceof Model)
-            {
+        } else {
+            if ($user instanceof Model) {
                 $user->refresh();
             }
 
             $email = $user->email;
         }
 
-        if (!$email)
-        {
+        if (!$email) {
             return;
         }