+++ src/EventListener/HookListener.php
@@ -1,36 +1,42 @@
 <?php
-/**
- * Contao Open Source CMS
+
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
  *
- * Copyright (c) 2018 Heimrich & Hannot GmbH
- *
- * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
- * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ * @license LGPL-3.0-or-later
  */
 
-
 namespace HeimrichHannot\Email2UsernameBundle\EventListener;
-
 
 use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
 use Contao\Database;
 use Contao\Input;
+use Contao\System;
 use Contao\Validator;
 
 class HookListener
 {
     /**
+     * @var bool
+     */
+    private $enabled = true;
+
+    /**
      * @var ContaoFrameworkInterface
      */
     private $framework;
 
-    public function __construct(ContaoFrameworkInterface $framework)
+    public function __construct(array $bundleConfig, ContaoFrameworkInterface $framework)
     {
+        if (isset($bundleConfig['user']) && true !== $bundleConfig['user']) {
+            $this->enabled = false;
+        }
+
         $this->framework = $framework;
     }
 
     /**
-     * This Hook provides case-insensitive contao-login by email usernames
+     * This Hook provides case-insensitive contao-login by email usernames.
      *
      * RFC 5321, section-2.3.11 says that email addresses should be treated as case-insensitive
      *
@@ -42,38 +48,55 @@ class HookListener
      */
     public function onImportUser($strUser, $strPassword, $strTable)
     {
-        if (!$this->framework->getAdapter(Validator::class)->isEmail($strUser))
-        {
+        if (!$this->enabled || !$this->framework->getAdapter(Validator::class)->isEmail($strUser)) {
             return false;
         }
 
-        switch ($strTable)
-        {
+        switch ($strTable) {
             case 'tl_member':
                 $objMember = $this->framework->createInstance(Database::class)->prepare('SELECT * from tl_member WHERE lower(username) = ?')->limit(1)->execute($strUser);
 
-                if ($objMember->numRows > 0)
-                {
+                if ($objMember->numRows > 0) {
                     // set post user name to the users username
                     $this->framework->getAdapter(Input::class)->setPost('username', $objMember->username);
 
                     return true;
                 }
+
                 break;
+
             case
             'tl_user':
                 $objUser = $this->framework->createInstance(Database::class)->prepare('SELECT * from tl_user WHERE lower(username) = ?')->limit(1)->execute($strUser);
 
-                if ($objUser->numRows > 0)
-                {
+                if ($objUser->numRows > 0) {
                     // set post user name to the users username
                     $this->framework->getAdapter(Input::class)->setPost('username', $objUser->username);
 
                     return true;
                 }
+
                 break;
         }
 
         return false;
+    }
+
+    public function onCreateNewUser($id, $data, $module)
+    {
+        if (!System::getContainer()->get('huh.utils.container')->isFrontend()) {
+            return;
+        }
+
+        if (!$this->enabled || !$module->reg_allowLogin || null === ($member = System::getContainer()->get('huh.utils.model')->findModelInstanceByPk('tl_member', $id))) {
+            return;
+        }
+
+        if (!$this->framework->getAdapter(Validator::class)->isEmail($member->email)) {
+            return;
+        }
+
+        $member->username = strtolower($member->email);
+        $member->save();
     }
 }