+++ src/Resources/contao/config/config.php
@@ -3,5 +3,5 @@
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['importUser']['huh_email2username'] = [
-    \HeimrichHannot\Email2UsernameBundle\EventListener\HookListener::class, 'onImportUser'];
+$GLOBALS['TL_HOOKS']['importUser']['huh_email2username'] = [\HeimrichHannot\Email2UsernameBundle\EventListener\HookListener::class, 'onImportUser'];
+$GLOBALS['TL_HOOKS']['createNewUser']['huh_email2username'] = [\HeimrichHannot\Email2UsernameBundle\EventListener\HookListener::class, 'onCreateNewUser'];